python cleanup, fix CI
