+++ src/tests/cucumber/features/steps/output_utils.py
@@ -1,3 +1,5 @@
+# Antares outputs parsing
+
 import os
 import time
 import datetime
